+++ frontend/src/assets/locales/en/organization.json
@@ -26,14 +26,18 @@
   "earlyIssuanceLabel": "Enable early issuance reporting",
   "edrmsLabel": "Organization profile, EDRMS record # (optional)",
   "serviceAddrLabel": "Address for service (postal address)",
-  "streetAddrLabel": "Street address / PO box",
+  "streetAddrLabel": "Street address/PO box",
   "addrOthLabel": "Address other (optional)",
   "cityLabel": "City",
   "provinceLabel": "Province",
-  "poLabel": "Postal / ZIP code",
+  "provinceStateLabel": "Province/State",
+  "poLabel": "Postal code",
+  "poZipLabel": "Postal code/ZIP code",
   "cntryLabel": "Country",
-  "bcAddrLabel": "Address in B.C. (at which records are maintained)",
-  "bcAddrLabelShort": "Address in B.C. (at which records are maintained)",
+  "recordsAddrGuide": "Enter full address including postal code (if different than address for service)",
+  "bcAddrLabel": "Head office (optional)",
+  "bcAddrLabelShort": "Head address (optional)",
+  "bcRecordLabel": "Address in B.C. where records are maintained (optional)",
   "sameAddrLabel": "Same as address for service",
   "contactMsg": "to update address information.",
   "usersLabel": "Users",
